[FIX #13] provide boolean to print with watermark logo or not
For now, there are boolean mentioned in program form view
and syllabus form view. It can be checked if we want syllabus
content with watermark else should be unchecked. But there is
as issue that, if one of the record the boolean is true and at
another record it might be true too. So for that, before print
we must double check the boolean as required even if the boolean
seems true/false as per needed.